+++ HISTORY.rst
@@ -8,6 +8,8 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 - #340: [Windows] Process.open_files() no longer hangs. Instead it uses a
   thred which times out and skips the file handle in case it's taking too long
   to be retrieved.  (patch by Jeff Tang, PR #597)
+- #627: [Windows] Process.name() no longer raises AccessDenied for pids owned
+  by another user.
 - #636: [Windows] Process.memory_info() raise AccessDenied.
 - #637: [UNIX] raise exception if trying to send signal to Process PID 0 as it
   will affect os.getpid()'s process group instead of PID 0.